auto-docs-archiver-back/requirements.txt
boto3==1.10.45
botocore==1.13.45
Click==7.0
docutils==0.15.2
Flask==1.1.1
Flask-Cors==3.0.8
Flask-JWT-Extended==3.24.1
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
joblib==0.14.0
MarkupSafe==1.1.1
nltk==3.4.5
numpy==1.17.4
pandas==0.25.3
Pillow==6.2.1
PyJWT==1.7.1
pymongo==3.9.0
pystempel==1.1.0
pytesseract==0.3.0
python-dateutil==2.8.1
pytz==2019.3
s3transfer==0.2.1
scikit-learn==0.21.3
scipy==1.3.3
six==1.13.0
sortedcontainers==2.1.0
tqdm==4.40.1
urllib3==1.25.7
Werkzeug==0.16.0
